Create Alice and bob's silly game.java
Input:5
Output:Alice
Explanation:We are given n=5 , so set={1,2,3,4,5} Alice chooses the prime number =2 and deletes the numbers2 and 4  and from the set, which becomes set={1,3,5}. Now there are two primes left,3 and 5  and . Bob can remove either prime from the set, and then Alice can remove the remaining prime. Because Bob is left without a final move, Alice will always win. Thus, we print Alice on a new line.