Create function of Jumping on the clouds by hackerrank.java
Emma is playing a new mobile game that starts with consecutively numbered clouds. Some of the clouds are thunderheads and others are cumulus. She can jump on any cumulus cloud having a number that is equal to the number of the current cloud plus 1 or 2 . She must avoid the thunderheads. Determine the minimum number of jumps it will take Emma to jump from her starting postion to the last cloud. It is always possible to win the game.
or each game, Emma will get an array of clouds numbered  if they are safe or  if they must be avoided. For example, c=[0,1,0,0,0,1,0] indexed from 0...6. The number on each cloud is its index in the list so she must avoid the clouds at indexes  1and5 . She could follow the following two paths: 0->2->4->6 or . The first path takes  jumps while the second takes 4 .
Sample Input:
7
0 0 1 0 0 1 0
Sample Output:4